Move `HttpAndCache` to `lib.rs`.
